Clippy fixes and refactor
